+++ string_calculator/implementation/src/main/java/pl/tzch96/string_calculator/Functions.java
@@ -1,8 +1,0 @@
-package pl.tzch96.string_calculator;
-
-public class Functions {
-
-    public static Double sqrt(Double x) {
-        return Math.sqrt(x);
-    }
-}