+++ string_calculator/main/src/main/java/pl/tzch96/string_calculator/App.java
@@ -3,10 +3,13 @@ package pl.tzch96.string_calculator;
 public class App 
 {
 
-    public static void main(String[] args)
-    {
+    public static void main(String[] args) {
         Calculator calculator = new Calculator();
 
         calculator.start();
+    }
+
+    public static void loadPlugins() {
+
     }
 }