+++ string_calculator/run.sh
@@ -1,5 +1,2 @@
-#FILE=./main/target/main-0.2.0-jar-with-dependencies.jar
-#if [ ! -f "$FILE" ]; then
-  mvn package
-#fi
-java -jar main/target/main-0.2.0-jar-with-dependencies.jar
+mvn clean install
+java -jar main/target/main-0.3.0-jar-with-dependencies.jar